Don't collect id-to-volume mapping until pages have come out of SheetDataSource
Initially, for pages in the volume body, the id-to-volume mapping was
done in VolumeProvider, but since the previous commit it was done also
in PageSequenceBuilder#nextPage, so in two places.

For pages in pre-/post-content, the id-to-volume mapping was only done
in PageSequenceBuilder#nextPage.

This commit moves the logic to one place, in VolumeProvider. Apart
from simplifying the code, this also fixes an issue in the regression
test suite. I have no proper explanation for it.